Sat, Jan 7, 2023, 7:37 PM +00:00

+++ readability/readability.c
@@ -27,8 +27,8 @@ int main(void)
             sentences++;
         }
     }
-    float L = letters / words * 100.0;
-    float S = sentences / words * 100.0;
+    float L = (float) letters / (float) words * 100;
+    float S = (float) sentences / (float) words * 100;
 
     int index = round(0.0588 * L - 0.296 * S - 15.8);
